requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
Django==5.1.6
django-allauth==65.4.1
django-jazzmin==3.0.1
django-widget-tweaks==1.5.0
idna==3.10
gunicorn
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
user-agents==2.2.0
whitenoise==6.9.0
